moved tile_button to new cpp
and made header
