+++ usr/Scores_list_10.txt
@@ -4,27 +4,29 @@ YAY 90
 
 YAY 80
 
-YAY 20
+YAY 80
 
 YAY 20
 
 YAY 20
 
 YAY 20
+
+YAY 20
+
+YAY 10
+
+YAY 10
+
+-- 10
+
+-- 10
+
+-- 10
+
+-- 10
+
+-- 10
 
 SY 10
-
--- 10
-
--- 10
-
--- 10
-
--- 10
-
--- 10
-
-YAY 10
-
-YAY 10
